+++ app/role-select/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import RoleSelect from "@/components/role-os/RoleSelect";
+
+export default function RoleSelectPage() {
+  return <RoleSelect />;
+}